refactor: generateLottos 로직 수정
